requirements.txt
Adafruit-Blinka==5.9.2
adafruit-circuitpython-neopixel==6.0.0
adafruit-circuitpython-pypixelbuf==2.2.0
Adafruit-PlatformDetect==2.23.0
Adafruit-PureIO==1.1.8
pkg-resources==0.0.0
pyftdi==0.52.0
pyserial==3.5
pyusb==1.1.0
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
sysv-ipc==1.0.1
